# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition"
  },
...
  "paths" : {
...
    "/v1/consent" : {
      "post" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "capture",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/consent/withdraw" : {
      "post" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "withdraw",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent-manager/{registrationId}/withdraw" : {
      "post" : {
        "tags" : [ "Consent Manager" ],
        "summary" : "Relay a withdrawal made at a Consent Manager",
        "operationId" : "withdraw_1",
        "parameters" : [ {
          "name" : "registrationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RelayWithdrawRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/consent-manager/{registrationId}/grant" : {
      "post" : {
        "tags" : [ "Consent Manager" ],
        "summary" : "Relay a grant made at a Consent Manager",
        "operationId" : "grant",
        "parameters" : [ {
          "name" : "registrationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RelayGrantRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelayResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rights/summary" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TypeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/receipts/{receiptId}" : {
      "get" : {
        "tags" : [ "receipt-controller" ],
        "operationId" : "receipt",
        "parameters" : [ {
          "name" : "receiptId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentReceipt"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/provenance/summary" : {
      "get" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "summary_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SourceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent/{entityId}/{subjectId}/receipt" : {
      "get" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "receipt_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentReceipt"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent-manager/{registrationId}/subjects/{cmSubjectRef}/record" : {
      "get" : {
        "tags" : [ "Consent Manager" ],
        "summary" : "The consent record UDS holds for a principal, as an ISO 27560 receipt",
        "operationId" : "record_1",
        "parameters" : [ {
          "name" : "registrationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "cmSubjectRef",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentReceipt"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CaptureRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "subject" : {
            "$ref" : "#/components/schemas/SubjectRef"
          },
          "alsoKnownAs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubjectRef"
            }
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "captureMethod" : {
            "type" : "string",
            "enum" : [ "CHECKBOX_OPT_IN", "CLICK_THROUGH", "DOUBLE_OPT_IN", "OTP_VERIFIED", "VERBAL_RECORDED", "WET_SIGNATURE", "BIOMETRIC_ENROLMENT", "PARENTAL_VERIFIED", "RELAYED_BY_CONSENT_MANAGER", "IMPORTED_WITH_PROVENANCE", "INFERRED_FROM_RELATIONSHIP", "NOT_APPLICABLE" ]
          },
          "actorType" : {
            "type" : "string",
            "enum" : [ "SUBJECT", "PARENT_GUARDIAN", "ADMIN", "AGENT", "SYSTEM", "IMPORT", "REGULATOR", "CONSENT_MANAGER" ]
          },
          "actorId" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "choices" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PurposeChoiceDto"
            },
            "minItems" : 1
          },
          "rejectAllOffered" : {
            "type" : "boolean"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "evidenceRef" : {
            "type" : "string"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "guardianVerification" : {
            "$ref" : "#/components/schemas/GuardianVerificationDto"
          }
        },
        "required" : [ "actorType", "captureMethod", "choices", "entityId", "jurisdiction", "languageTag" ]
      },
      "GuardianVerificationDto" : {
        "type" : "object",
        "properties" : {
          "method" : {
            "type" : "string",
            "enum" : [ "EXISTING_VERIFIED_ACCOUNT", "DIGILOCKER_VIRTUAL_TOKEN", "OTHER_DOCUMENTED" ]
          },
          "reference" : {
            "type" : "string",
            "minLength" : 1
          },
          "verifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verifiedBy" : {
            "type" : "string"
          }
        },
        "required" : [ "method", "reference", "verifiedAt" ]
      },
      "PurposeChoiceDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "granted" : {
            "type" : "boolean"
          },
          "preTicked" : {
            "type" : "boolean"
          },
          "separateAction" : {
            "type" : "boolean"
          }
        },
        "required" : [ "purposeCode" ]
      },
      "SubjectRef" : {
        "type" : "object",
        "properties" : {
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "value" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "identifierType", "value" ]
      },
      "CaptureResponse" : {
        "type" : "object",
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "subjectId" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventSummary"
            }
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ViolationDto"
            }
          }
        }
      },
      "EventSummary" : {
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "string"
          },
          "purposeCode" : {
            "type" : "string"
          },
          "purposeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "eventType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "GRANTED", "DENIED", "WITHDRAWN", "NOT_ASKED", "EXPIRED", "INVALIDATED", "PENDING_SYNC", "CONFLICTED", "UNKNOWN" ]
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "sequenceNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventHash" : {
            "type" : "string"
          }
        }
      },
...
      "WithdrawRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "subject" : {
            "$ref" : "#/components/schemas/SubjectRef"
          },
          "purposeCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "minItems" : 1
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "actorType" : {
            "type" : "string",
            "enum" : [ "SUBJECT", "PARENT_GUARDIAN", "ADMIN", "AGENT", "SYSTEM", "IMPORT", "REGULATOR", "CONSENT_MANAGER" ]
          },
          "actorId" : {
            "type" : "string"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        },
        "required" : [ "actorType", "entityId", "jurisdiction", "purposeCodes" ]
      },
...
      "RelayWithdrawRequest" : {
        "type" : "object",
        "properties" : {
          "cmSubjectRef" : {
            "type" : "string",
            "minLength" : 1
          },
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "identifierValue" : {
            "type" : "string"
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "purposeCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "minItems" : 1
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          }
        },
        "required" : [ "cmSubjectRef", "entityId", "jurisdiction", "purposeCodes" ]
      },
      "RelayResponse" : {
        "type" : "object",
        "properties" : {
          "registrationId" : {
            "type" : "string"
          },
          "cmSubjectRef" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventSummary"
            }
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ViolationDto"
            }
          }
        }
      },
      "RelayGrantRequest" : {
        "type" : "object",
        "properties" : {
          "cmSubjectRef" : {
            "type" : "string",
            "minLength" : 1
          },
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "identifierValue" : {
            "type" : "string"
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "choices" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PurposeChoiceDto"
            },
            "minItems" : 1
          },
          "rejectAllOffered" : {
            "type" : "boolean"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "evidenceRef" : {
            "type" : "string"
          }
        },
        "required" : [ "choices", "cmSubjectRef", "entityId", "jurisdiction", "languageTag" ]
      },
...
      "TypeSummary" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "open" : {
            "type" : "integer",
            "format" : "int64"
          },
          "overdue" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closed" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closedLate" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "ConsentReceipt" : {
        "type" : "object",
        "properties" : {
          "schemaVersion" : {
            "type" : "string"
          },
          "receiptId" : {
            "type" : "string"
          },
          "consentRecordId" : {
            "type" : "string"
          },
          "consentManagerRegistrationId" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fiduciaryName" : {
            "type" : "string"
          },
          "fiduciaryId" : {
            "type" : "string"
          },
          "dpoContact" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Entry"
            }
          },
          "withdrawalUri" : {
            "type" : "string"
          },
          "rightsUri" : {
            "type" : "string"
          },
          "grievanceUri" : {
            "type" : "string"
          },
          "evidenceHash" : {
            "type" : "string"
          },
          "parentalVerification" : {
            "type" : "string",
            "enum" : [ "EXISTING_VERIFIED_ACCOUNT", "DIGILOCKER_VIRTUAL_TOKEN", "OTHER_DOCUMENTED" ]
          }
        }
      },
...
      "SourceSummary" : {
        "type" : "object",
        "properties" : {
          "sourceType" : {
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "quarantined" : {
            "type" : "integer",
            "format" : "int64"
          },
          "contactable" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}